feat: extract general request mapping helpers
The common gateway provides basic request functionality that is common to REST and gRPC alike.
